requirements.txt
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.2
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dungeon-maps==0.0.3a1
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
ftfy==6.1.1
gitdb==4.0.9
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1676637751542/work
gym==0.26.2
gym-notices==0.0.8
-e git+https://github.com/facebookresearch/habitat-lab.git@bc85d0961cef3b4a08bc9263869606109fb6ff0a#egg=habitat
habitat-sim==0.2.1
idna==3.4
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1686552404404/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1673483481485/work
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
llvmlite @ file:///croot/llvmlite_1683555126642/work
magnum @ file:///opt/conda/envs/py36/conda-bld/habitat-sim_1627788414467/work/build/deps/magnum-bindings/src/python
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1685544882857/work
numpy==1.21.6
numpy-quaternion @ file:///home/conda/feedstock_root/build_artifacts/quaternion_1676352986710/work
omegaconf==2.3.0
open3d-python==0.3.0.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2023.6.3
requests==2.28.1
rich==12.6.0
scikit-learn==1.0.2
scipy==1.9.1
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.15.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu116
torch-scatter==2.1.1
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
vis==0.0.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
yacs==0.1.8
yapf==0.32.0
zipp==3.10.0
